Small emprove of readability
